backend/requirements.txt
# LangChain core framework
langchain
langchain-openai
langchain-chroma
langchain-community

# LangGraph for stateful workflows
langgraph

# Document processing
pypdf
unstructured==0.18.15
markdown==3.9

# Environment configuration
python-dotenv

